Implement add preset feature
